backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
yfinance==0.2.51
pandas==2.2.3
pandas-ta==0.3.14b0
openai==1.65.0
numpy==2.2.3
python-dotenv==1.0.1
httpx==0.28.1
